Use opctx_alloc in start saga to query boundary switches
Finding boundary switches requires fleet access; use the special instance-
allocation context to do this (this context was already used for sled lookups
earlier in the relevant saga step). This was not previously a problem because
NAT entries were always created by the instance create saga, which got a
boundary switch list from its creator that the creator obtained by querying
using the instance allocation context.

Update the instance-in-silo integration test to make sure that its instances
do actually start. This required a somewhat heavy lift: test helper functions
like `instance_simulate` use `OpContext::for_tests` to get an operation context
to use to query the database, but this context is associated with a user in the
default silo. To get around this, the test hacks up a "background" test context
with the user it created in its test silo. This seems to work OK, but I'm open
to alternative ways of doing this.

Tested with cargo tests.