Combine host and trampoline OS images into one CI job
There are two OS build jobs that each check out and build many
repositories in order to build a host image. By combining them
into one job, we only pay the cost of this checkout and build
phase once. It doesn't actually save a lot of wall time because
the two jobs run in parallel, but it saves starting a VM and
about 20 minutes of compute time.