Create zone bundles from ZFS snapshots (#4225)
- Fixes #4010
- Previously, we copied log files directly out of their original
locations, which meant we contended with several other components:
`logadm` rotating the log file; the log archiver moving the to
longer-term storage; and the program writing to the file itself. This
commit changes the operation of the bundler, to first create a ZFS
snapshot of the filesystem(s) containing the log files, clone them, and
then copy files out of the clones. We destroy those clones / snapshots
after completing, and when the sled-agent starts to help with
crash-safety.