Implement image deletion
Also, make sure that multiple volume construction requests can only be
copied for read-only Regions: the `region` table has a column for the
volume that owns it, and this means there's a one to one mapping for the
volume that references the region read-write. Ensure this in
`volume_checkout_randomize_ids` otherwise multiple volumes could
reference the same regions and this could lead to accounting / cleanup
issues.

Fixes #3033