plumb customer-configured DNS into instances (#4265)
Currently [OPTE hardcodes DNS servers to
8.8.8.8](https://github.com/oxidecomputer/opte/issues/390), and we would
prefer to not do this.

When setting up the rack, customers configure DNS servers. Currently
this is added to Nexus's deployment configuration, so that Nexus can
resolve external domain names for SAML. This change plumbs these DNS
servers through to sled-agent and OPTE.

For non-instance OPTE ports, `DhcpCfg::default()` is used, which is
equivalent to "no hostname, no domain name, no DNS servers, no search
domains".

This is not the final state of the DHCP work we want to do; in talking
with @rmustacc and @rcgoodfellow we agree that implementing this minimal
DHCP option set is the correct thing to do urgently; and that we still
need to design the broader inter-instance networking picture (whether we
want an instance-facing recursive resolver on the rack and whether it
should resolve domain names for other instances with VPC IPs, what those
hostnames and search domains should be, and how customers can modify
these settings).