[wicketd] Accept TUF repos with RoT archives signed with different keys (#4289)
As of this PR, wicketd will (a) accept TUF repos containing multiple RoT
archives for the same board target (e.g., multiple gimlet RoT images),
and when performing a mupdate, it will ask the RoT for its
currently-active CMPA and CFPA pages and search for an RoT archive that
matches.

After this is deployed to all fielded systems, we'll be able to drop the
`-rot-staging-dev` and `-prod-rel` TUF repos from CI, and only build a
single TUF repo with all RoT images. This PR adds a new `-rot-all` TUF
repo publishing step but does not remove the old ones, as we'll need
them to update into this version of wicketd.