Make sure project_delete_snapshot does its job
`project_delete_snapshot`'s job is to soft-delete snapshots but due to
the combination of the time_deleted filter and `check_if_exists` it was
not doing this: `test_create_snapshot_record_idempotent` would fail if
there was a check that the snapshot was actually deleted.

This was causing a snapshot delete request to proceed for a snapshot
that was in state Creating, causing a case where a volume was hard
deleted before it had its associated resource accounting performed,
leading to regions associated with deleted resources that would never be
cleaned up.