Make sure project_delete_snapshot does its job (#4308)
`project_delete_snapshot`'s job is to soft-delete snapshots but due to
incorrectly using `check_if_exists` with `execute_async`, the state a
snapshot was in was not being validated. This was causing a snapshot
delete request to proceed for a snapshot that was in state Creating,
causing a case where a volume was hard deleted before it had its
associated resource accounting performed, leading to regions associated
with deleted resources that would never be cleaned up.

Fix this and ensure expected behaviour via a test.