[wicketd] Fall back to old behavior if SP is too old to support reading RoT CMPA/CFPA (#4326)
We hit this on rack3 but did not hit it on dogfood due to more frequent
updates to dogfood's SP/RoT. Prior to this PR, wicketd expected to be
able to ask an SP for its RoT's CMPA/CFPA pages, but if a rack is
jumping from the 1.0.2 release to current master, its SPs are too old to
understand that message. With this change, we will fall back to
wicketd's previous behavior of requiring exactly 1 RoT archive if we
fail to fetch the CMPA from the target component.