Initial work for replicated ClickHouse testing (#4149)
## Structural changes

Instead of spinning up a ClickHouse server per test, now only two
servers are spun up to run all tests against. One for single node mode
and another for replicated testing. Even though it may look odd to this
this now, there is a good reason for this. ClickHouse servers that have
been configured to be part of a cluster cannot be spun up in the same
server as single node servers. This means that if we spin up several
servers, the replicated test becomes incredibly flaky. It's important to
note that this set up will not live forever. Once we can remove single
node mode and start the replicated ClickHouse cluster as part of the
build process, the replicated tests will be able to run independently.

All tests now run sequentially against a single server. For correctness,
after each test the database is wiped out to give us a clean slate for
testing. This only adds a few more seconds (<10s) to the overall testing
time. Again, this will not be the final state of things, but necessary
to be able to test alongside replicated clusters until we
[remove](https://github.com/oxidecomputer/omicron/issues/4148) single
node functionality.

## Caveat

Github does some really weird things with the diffs on
oximeter/db/src/client.rs. Frankly, it looks like a mess. I would
probably checkout the branch itself to see the final state of the tests.

## Follow up work

Given the size of this PR, I've not added all of the tests to run
against the replicated cluster. There will be a follow up PR that does
this.

Related: https://github.com/oxidecomputer/omicron/issues/4001
Closes: https://github.com/oxidecomputer/omicron/issues/4037