[wicket] switch ComponentId Display impl to lowercase (#4358)
The current ComponentId Display impl is uppercase, which is ok for
wicket but we'd like to use. Lowercase is a better default, so use that.
(But PSC is an acronym so leave it as uppercase.)

I checked this by commenting out the `Display` impl (and the
`Into<Text>` impl), then inspecting all the places that errored out. I
changed all the spots that seemed to fit better with uppercase, and left
the ones that would work better with lowercase (typically log messages)
untouched.

The `Into<Text>` impl seemed unnecessary so I just removed it.