[oximeter] Port all single node ClickHouse tests to a replicated set up (#4372)
This is a follow up to
https://github.com/oxidecomputer/omicron/pull/4149 . In this commit all
single node tests are now run against a clustered set up. Additionally,
a few bugs on the replicated init SQL file have been fixed.

Note:

All of the individual field and measurement tests all run under a
single test now. The reason for this is test times. Each test needs to
wipe out it's database before the next test runs. In a replicated
installation, this means that all nodes must sync, which takes time.
Before I merged all the field/measurement tests into a single one, the
testing time for a replicated set up took about 10 minutes, which really
is too much. Now it takes ~3.5 minutes. It is still a lot, but only a
temporary measure until we can run these tests in parallel again. An
unintended bonus of this approach is that running the tests this way
means that you are testing that the data is being inserted in the
correct table.

Related: https://github.com/oxidecomputer/omicron/issues/4148
Related: https://github.com/oxidecomputer/omicron/issues/4001