Implement sled-agent API for adding a sled
This code largely reuses the existing code paths for starting sled-agent
with a small change to configured the bootstore as a learner node if
trust quorum is in use.

This code path does not attempt any retries, as it should be idempotent
as long as the same sled and rack UUIDs are used.This is a command
driven by an operator and we want to know right away if it succeeded or
not. Multiple sleds can be added by Nexus with individual calls, so that
individual errors can be returned without added complexity at the sled-
agent level. These requests can also be sent to different sled-agents,
since the sled-agent is just acting as a proxy to a remote
bootstrap agent.