sled-agent's Instance doesn't need PROPOLIS_PORT
The full SocketAddr is provided by nexus, where it's still hardcoded.
No sense in stripping the port off of that and then re-adding it here
(or so I think!)

Fixing nexus's side of PROPOLIS_PORT being hardcoded is likely to entail
a db schema change, but this comparitively small change will already make
it more possible to write tests in sled-agent.