[wicket] add rack-update clear (#4451)
As a followup to #4062, add a command to clear update state via the CLI.

To test this command, we make it have a `--message-format` option which
outputs JSON. This PR also adds support for reading that JSON, by moving
`ClearUpdateStateResponse` into `wicket-common`. Doing so meant that
`SpIdentifier` and `SpType` also moved into `wicket-common`. (We should
maybe unify this with the gateway types, but that isn't necessary for
this change.)

(Side note: I tried implementing `--message-format json` for
`rack-update start` and `attach` as well, but that's a somewhat bigger
project with dubious upside so I shelved it for now. Implementing
`clear` was easy.)