Include enums in `dbinit_equal_sum_of_all_up` test (#4472)
While working on #4261 I ~~ran into~~ created a situation where the
order of items in a DB enum matters because I want to sort by it. In
this case it's about specificity — the type is `silo` or `fleet` and I
want to sort by it to ensure I get the most "specific" thing first.
@zephraph noticed I had an order mismatch between my migration adding
the enum and the copy in `dbinit.sql`, and it wasn't caught by the
tests.

So, here I've added the list of user-defined enums to the set of things
we compare when we ask whether the result of our migrations matches the
contents of `dbinit.sql`. There were two existing enums that failed the
test because the copy in `dbinit.sql` was sorted alphabetically, but a
migration added a new item at the end of the list. Very easy to fix,
though now are enums aren't sorted nicely, which is sad.

The question is: should we enforce enum order? Seems like we might as
well if it doesn't cost us much. But should we write application logic
like #4261 that relies on enum order? In that case, at least, I have a
reasonable alternative because the list of things I'm sorting is at most
2 long, so I can easily get the whole list and pick the element I want
in app code instead of doing an `order by` + `limit 1` in SQL. It is
made quite clear in a comment, but it still feels a little dicey.