Add offline schema updates for ClickHouse (#4394)
- Some cleanup around issuing multiple SQL statements from a file
- Create directory structure for storing schema updates modeled after
CRDB up.sql files, but using integer versions, and move all existing SQL
into version 2
- Add version 3, which fixes
https://github.com/oxidecomputer/omicron/issues/4369, but does not apply
it yet
- Add methods in the client for listing, reading, and applying one or
more updates to the oximeter database from the upgrade files
- Add tests for upgrade application
- Add `clickhouse-schema-updater` binary for running them on demand
- Modify `oximeter-collector` to _not_ wipe / reinit the DB on startup
if the version has change, but instead wait for the version to be equal
to what it is compiled against. This relies on updates from the
developer being applied before `oximeter` will continue.