Improve wtmp record mangling.
This fixes two problems with the current wtmp/utmp record mangling.
First it is not idempotent, the system's notion of boot time is
set to the current time of day whenever sled agent confirms that
time is synchronised. Secondly, this is only approximate even for the
first sled agent start, but it's plain wrong if sled agent restarts
later.

In conjunction with changes to stlouis, the `tmpx` utility is now
able to process all zones itself, and uses the true system boot
time for each zone when updating records.

Fixes: https://github.com/oxidecomputer/omicron/issues/3514