[nexus] List all uninitialized sleds (#4504)
As part of adding a sled to an already initialized rack, we need a way
for operators to be able to list sleds that are not part of a rack. This
PR adds an external nexus endpoint for doing just that. Like the `sleds`
endpoint, this endpoint is not tied to a rack.

The way this works is by looking at the SPs in the latest inventory
collection and finding all the Baseboards that are not in the `sled`
table in CRDB.

A follow up commit will allow an operator to add uninitialized sleds to
a rack with a new external nexus endpoint.