[wicketd] Fix infinite retry of trampoline upload to MGS
If the trampoline phase 2 image that we upload to MGS is bad (e.g.,
because it's fake data that doesn't contain a valid OS header), we would
previously retry the doomed upload forever with exponential backoff.
With this change, we instead try 3 times (which should be more than
sufficient for any blips, as we're talking to MGS in our same zone over
localhost) and then fail the update.

Now that the simulated SP and RoT can accept updates, without this
change we landed in an infinite loop in wicketd's update test, as it was
(infinitely) waiting for the update to fail, but the update was stuck in
the infinite upload retry.