Add a producer kind to oximeter metric producers (#4497)
- Adds the `kind` enum to metric producer information, including DB
schema, model, and various client parameter types. This records the
supported types of metric producers, and is intended to aid debugging
and future work around updates and instance lifecycle management.
- Add schema update files which create the DB enum type and add it as a
column to the `metric_producer` table. This currently _drops_ the
existing table and recreates it with the new column, rather than adding
the column using `ALTER TABLE`. That is intended to remove old entries
in bulk, since nothing previously removed the records for Propolis
servers when their instance was stopped.

This is the initial PR in a sequence that will eventually make this
field _required_ in both the database and API requests. As there are
consumers of this API outside of the Omicron repository, this field
needs to start as optional, to avoid introducing a commit with
incompatible clients.