[nexus-db-model] separate out SledUpdate from Sled (#4533)
`Sled` consists of several columns that aren't controlled by sled-agent, and we
end up in this weird place where we have `Sled` instances that don't reflect
reality. I'm working on adding a `provision_state` column which is controlled
by the operator, and again for which sled-agent doesn't know.

Clean this up by defining a new struct, `SledUpdate`, which only contains the
columns sled-agent knows about. The other columns get defaults when
`into_insertable` is called.