Various volume management fixes (#4410)
This commit bundles up a few fixes related to volume management:

- `find_deleted_volume_regions` now returns the `Option<RegionSnapshot>`
that resulted from the `left_join` in the query for freed regions. This
is then consulted to see if sending a DELETE for that region is safe: if
the `Option` is `Some`, then the region snapshot has not been deleted
yet, and sending a region DELETE will surely result in a `must delete
snapshots first` error from the corresponding Crucible agent.

- Fix a few typos in nexus/src/app/sagas/common_storage.rs

- Nexus now waits for the Agent's `zfs destroy` of a snapshot to take
place. Otherwise if illumos doesn't immediately remove the snapshot it
may be returned by a subsequent `zfs list` later.

- Either `decrease_crucible_resource_count_and_soft_delete_volume` or
`volume_hard_delete` should be called when unwinding a saga, calling
both is not required.

- In the snapshot deletion saga, use `append_parallel` for the two
volume delete sub sagas: in order to _not_ orphan Crucible resources,
it's important that both volumes be soft deleted, and that failing to
delete one volume's Crucible resources does not cause the other to _not_
be soft deleted.

- Also fix a very confusing typo when building the destination volume
delete sub saga.