[update-engine] fix GroupDisplayStats to avoid integer underflow (#4561)
This could happen if an empty `EventReport` is passed in -- in that case we'd
transition to `Running` but return `NotStarted`.

Fix this by not transitioning `self.kind` to `Running` if we're going to return
`NotStarted`. This does bloat up the code a little but I think is clearer
overall.

Thanks to @jgallagher for all the help debugging this!

Also clean up some related logic and add tests.

Fixes #4507.