Detect and un-delete phantom disks (#4547)
A "phantom" disk is a disk where the disk delete saga has run for it but
unwound: this leaves that disk soft-deleted, but the resources and
accounting for that disk remain. Users cannot request that the disk be
deleted again, and it remains a phantom.

There are two fixes for this:

1. Change the disk delete saga to undo the disk's soft delete and set
the disk to faulted during an unwind. This way, users can request that
disks be deleted repeatedly until it works.

2. Create a background task that detects these phantom disks and does
the same thing: un-delete them and set them to faulted.

This requires adding an index on `id` to the `disk` table, so the schema
is bumped to 12.0.1.

Fixes oxidecomputer/customer-support#58.