Support missing samples in oximeter (#4552)
- Add a `Datum::Missing` and `MissingDatum`, which records the intended
datum type and an optional start time for a sample which could not be
produced.
- Database upgrades which make all scalar datum columns Nullable. Array
fields are _not_ made Nullable, since ClickHouse doesn't support
composite types like arrays inside a Nullable wrapper type. The empty
array is used as a sentinel, which is OK since we can't have zero-length
array histograms in Oximeter. Add a test which will fail if we ever
change that.
- Rework database serialization to handle Nullable types or empty
arrays. This uses a new helper trait to convert a NULL (which has no
type information) to the intended datum type, or an empty array to a
histogram.
- Add a test for each measurement type that we can recover a missing
sample of that type -- NULLs for scalar values and empty arrays for
histograms.