[sled-agent] add preliminary "write boot disk OS" http endpoints (#4633)
This PR adds three new endpoints to `sled-agent`:

* `POST /boot-disk/{boot_disk}/os/write` to start a new update to one of
the two boot disks
* `GET /boot-disk/{boot_disk}/os/write/status` to get the status of the
most-recently-started update to the specified boot disk
* `DELETE /boot-disk/{boot_disk}/os/write/status/{update_id}` to clear the
  status of a previous update

The actual drive-writing-machinery is extracted from `installinator`
into `installinator-common`, which is now a new dependency of
`sled-agent`. The bulk of the changes in this PR center around being
able to run that drive-write from within the context of the pair of
dropshot endpoints above, plus a bit of glue for unit testing with
"in-memory disks".