[sled-agent/sled-hardware] remove serial_test dependency (#4656)
serial_test is a proc macro which ensures that tests run serially rather than
in parallel. However, as documented at https://nexte.st/book/test-groups.html,
serial_test doesn't actually work with nextest. So go ahead and remove it as a
dependency.

The next question is: do we need to replace it with nextest's test groups? It
turns out that the answer is "no". The only uses of serial_test in our codebase
were to enable testing against a mocked free function (if multiple tests would
call that function concurrently, the mocking infrastructure would get
confused). However, that is only an issue if multiple tests are run within the
same process. Nextest's execution model is process-per-test, which means that
this isn't an issue at all.

(This is also hinted at by the fact that `serial_test` has effectively been
inoperative for months, yet we've had no issues with these tests.)