add repro test for 4266
Add a start saga integration test that shows provisioning counters being leaked
when an instance fails to start in the ensure_running step. This requires some
simulated sled agent changes to (a) provide a means to inject this sort of
failure and (b) not make an "unregister instance" operation send updates to
sled agent through the asynchronous notification channel (which happens to clean
up the provisioning counters on its own in a way that makes the test pass).