Oximeter self-stat tests don't need to be based on time (#4670)
- Fixes #4657
- Oximeter self stat tests run a few collections, and verify that the
relevant counters keep track of those. Previously, that was based on
_time_, advancing time forward until the expected number of collections
actually occur. There's nothing really guaranteeing that those
collections do occur, since the tasks may not run exactly when we
expect. This commit updates the tests so that the producer explicitly
updates a shared counter every time they receive a request. The tests
then reduce to asserting this counter equals the self-stat counter.